update db definitions for data types; update seed files to conform to new definitions

+++ server/db/models/item.js
@@ -24,10 +24,10 @@ const Item = db.define('item', {
     }
   },
   price: {
-    type: Sequelize.DECIMAL,
+    type: Sequelize.INTEGER,
     allowNull: false,
     validate: {
-      min: 0.0
+      min: 0
     }
   },
   color: {
